gcs-to-bigquery/requirements.txt
google-cloud-storage==2.18.0
google-cloud-bigquery==3.25.0
flask==3.0.3
gunicorn==23.0.0